test: move sample test to separate class
